+++ system/src/apps/_module.mjs
@@ -1,0 +1,2 @@
+export {default as CombatTracker2d20} from "./CombatTracker2d20.mjs";
+export {default as MomentumTracker} from "./MomentumTracker.mjs";