+++ system/src/documents/_module.mjs
@@ -1,0 +1,3 @@
+export {default as ACActor} from "./ACActor.mjs";
+export {default as ACItem} from "./ACItem.mjs";
+export {default as Combat2d20} from "./Combat2d20.mjs";