+++ system/src/hooks/renderChatMessageHook.mjs
@@ -1,0 +1,9 @@
+import { ACChat } from "../system/ACChat.mjs";
+
+export const renderChatMessageHook = {
+	attach: () => {
+		ac2d20.logger.debug("Attaching renderChatMessage hook");
+
+		Hooks.on("renderChatMessage", ACChat.onRenderChatMessage);
+	},
+};