+++ system/src/sheets/ACNPCSheet.mjs
@@ -1,12 +1,10 @@
-// import { AC2D20 } from "../helpers/config.mjs";
-// import { onManageActiveEffect, prepareActiveEffectCategories } from "../helpers/effects.mjs";
-import { ACActorSheet } from "./actor-sheet.mjs";
+import ACActorSheet from "./ACActorSheet.mjs";
 
 /**
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ACActorSheet}
  */
-export class ACNPCSheet extends ACActorSheet {
+export default class ACNPCSheet extends ACActorSheet {
 
 	/** @override */
 	static get defaultOptions() {
@@ -15,7 +13,11 @@ export class ACNPCSheet extends ACActorSheet {
 			template: "systems/ac2d20/templates/actor/npc-sheet.hbs",
 			width: 550,
 			height: 780,
-			tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "abilities" }],
+			tabs: [{
+				navSelector: ".sheet-tabs",
+				contentSelector: ".sheet-body",
+				initial: "abilities",
+			}],
 		});
 	}
 }